+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Version 0.14.0
+## Bug fixes
+- Fixed height not working in REST queries (#299)
 - Implemented blocked users checks (#298)
 
 # Version 0.13.0